Use mpfr::mpreal as arithmetic type, Implement settings, Load and save settings, Set default mpfr precision from settings, Degrade mpfr::mpreal to double in exprtk module and PySymbolTable instantiation routine (To be replaced with custom python type), Implement fraction test for mpfr::mpreal, Implement settings serialization, Degrade mpfr::mpreal to / from double when serializing / deserializing, Implement numberformatting for mpfr::mpreal

+++ src/cpp/pymodule/exprtkmodule.cpp
@@ -33,7 +33,7 @@ PyObject *evaluate(PyObject *self, PyObject *args) {
 
         PyObject *ret = PyTuple_New(2);
 
-        PyTuple_SetItem(ret, 0, PyFloat_FromDouble(value));
+        PyTuple_SetItem(ret, 0, PyFloat_FromDouble(value.toDouble()));
         PyTuple_SetItem(ret, 1, PySymbolTable::New(symTable));
 
         PySymbolTable::Cleanup(symTable);
